+++ auth-service/src/main/java/org/example/auth/service/PersonService.java
@@ -4,7 +4,7 @@ import org.example.auth.dto.mmtr.MmtrAuthResponseEmployeeDto;
 import org.example.grpc.PeopleService;
 
 public interface PersonService {
-    PeopleService.Person findPersonById(Integer personId);
+    PeopleService.Person findPersonById(Long personId);
 
     void savePersonIfNotExists(MmtrAuthResponseEmployeeDto authResponseEmployee);
 }
