+++ course-registration-service/src/main/java/org/example/courseregistration/exception/classes/course/CourseNotFoundException.java
@@ -1,7 +1,7 @@
 package org.example.courseregistration.exception.classes.course;
 
 public class CourseNotFoundException extends RuntimeException {
-    public CourseNotFoundException(Integer courseId) {
+    public CourseNotFoundException(Long courseId) {
         super("Система не обнаружила у себя в памяти курс " + courseId);
     }
 }
