+++ course-registration-service/src/main/java/org/example/courseregistration/repository/CourseRegistrationRequestRepository.java
@@ -9,24 +9,24 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface CourseRegistrationRequestRepository extends JpaRepository<CourseRegistrationRequest, Integer> {
+public interface CourseRegistrationRequestRepository extends JpaRepository<CourseRegistrationRequest, Long> {
     @Query("SELECT crr FROM CourseRegistrationRequest crr\n" +
             " WHERE crr.personId = :personId AND crr.status.status != 'ACCEPTED'")
-    Optional<CourseRegistrationRequest> findCourseRegistrationRequestByPersonIdAndStatus_NotAccepted(@Param("personId") Integer personId);
+    Optional<CourseRegistrationRequest> findCourseRegistrationRequestByPersonIdAndStatus_NotAccepted(@Param("personId") Long personId);
 
-    List<CourseRegistrationRequest> findCourseRegistrationRequestsByRequestIdIn(List<Integer> requestIds);
+    List<CourseRegistrationRequest> findCourseRegistrationRequestsByRequestIdIn(List<Long> requestIds);
 
     @Query(value = "SELECT crr FROM CourseRegistrationRequest crr\n" +
             "JOIN CourseRegistrationRequestKeeper crrk ON crrk.requestId = crr.requestId\n" +
             "WHERE crrk.keeperId = :keeperId AND crr.status.status = 'APPROVED' AND crrk.status.status = 'APPROVED'\n" +
             "ORDER BY crrk.responseDate")
-    List<CourseRegistrationRequest> findApprovedRequestsByKeeperId(@Param("keeperId") Integer keeperId);
+    List<CourseRegistrationRequest> findApprovedRequestsByKeeperId(@Param("keeperId") Long keeperId);
 
     @Query(value = "SELECT crrk FROM CourseRegistrationRequest crr\n" +
             "JOIN CourseRegistrationRequestKeeper crrk ON crrk.requestId = crr.requestId\n" +
             "WHERE crrk.keeperId IN :keeperIds AND crr.status.status = 'APPROVED' AND crrk.status.status = 'APPROVED'\n" +
             "ORDER BY crrk.responseDate")
-    List<CourseRegistrationRequestKeeper> findApprovedKeeperRequestsByKeeperIdIn(@Param("keeperIds") List<Integer> keeperIds);
+    List<CourseRegistrationRequestKeeper> findApprovedKeeperRequestsByKeeperIdIn(@Param("keeperIds") List<Long> keeperIds);
 
-    void deleteCourseRegistrationRequestsByCourseId(Integer courseId);
+    void deleteCourseRegistrationRequestsByCourseId(Long courseId);
 }
