+++ course-registration-service/src/main/java/org/example/courseregistration/repository/CourseRegistrationRequestStatusRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CourseRegistrationRequestStatusRepository extends JpaRepository<CourseRegistrationRequestStatus, Integer> {
+public interface CourseRegistrationRequestStatusRepository extends JpaRepository<CourseRegistrationRequestStatus, Long> {
     Optional<CourseRegistrationRequestStatus> findCourseRegistrationRequestStatusByStatus(CourseRegistrationRequestStatusType status);
 }
