+++ course-registration-service/src/main/java/org/example/courseregistration/repository/ExplorerRepository.java
@@ -7,9 +7,9 @@ import java.util.List;
 import java.util.Optional;
 
 public interface ExplorerRepository {
-    Optional<ExplorersService.Explorer> findExplorerByPersonIdAndGroup_CourseId(Integer personId, Integer courseId);
+    Optional<ExplorersService.Explorer> findExplorerByPersonIdAndGroup_CourseId(Long personId, Long courseId);
 
     void save(ExplorerCreateEvent explorer);
 
-    ExplorersService.ExplorersByPersonIdAndGroup_CourseIdInResponse findExplorersByPersonIdAndGroupCourseIdIn(Integer personId, List<Integer> courseIds);
+    ExplorersService.ExplorersByPersonIdAndGroup_CourseIdInResponse findExplorersByPersonIdAndGroupCourseIdIn(Long personId, List<Long> courseIds);
 }
