+++ course-registration-service/src/main/java/org/example/courseregistration/repository/GalaxyRepository.java
@@ -3,5 +3,5 @@ package org.example.courseregistration.repository;
 import org.example.courseregistration.dto.galaxy.GalaxyDto;
 
 public interface GalaxyRepository {
-    GalaxyDto findGalaxyBySystemId(Integer systemId);
+    GalaxyDto findGalaxyBySystemId(Long systemId);
 }
