+++ course-registration-service/src/main/java/org/example/courseregistration/service/CourseRegistrationRequestKeeperService.java
@@ -16,7 +16,7 @@ public interface CourseRegistrationRequestKeeperService {
 
     void closeRequestToOtherKeepers(CourseRegistrationRequest request);
 
-    List<CourseRegistrationRequestKeeper> findCourseRegistrationRequestKeepersByRequestId(Integer requestId);
+    List<CourseRegistrationRequestKeeper> findCourseRegistrationRequestKeepersByRequestId(Long requestId);
 
-    List<CourseRegistrationRequestKeeper> findProcessingCourseRegistrationRequestKeepersByKeeperIdIn(List<Integer> keeperIds);
+    List<CourseRegistrationRequestKeeper> findProcessingCourseRegistrationRequestKeepersByKeeperIdIn(List<Long> keeperIds);
 }
